Factor out AddressAndBoroughField.
